Server side auth-tokens
created AuthTokenEntity, AuthTokenService, AuthTokenRepository. New token will be created in auth_tokens table everytime a new user gets created or a user logs in. This token is sent back in response body to the client's browser where it will be stored either in cookies or local-storage and in every subsequent request will contain this token for authenticating user.